+++ src/js/components/refs.js
@@ -1,0 +1,3 @@
+export default {
+  galleryContainer: document.getElementById('js-gallery'),
+};
